move convertToReact to related page lib folder
