+++ lib/pages/posts/[slug].ts
@@ -1,7 +1,10 @@
+import { createElement, Fragment } from 'react'
+import rehypeReact from 'rehype-react'
+import { ComponentsWithoutNodeOptions } from 'rehype-react/lib/complex-types'
+import { unified } from 'unified'
 import { HastTree, Post, PostMeta } from 'lib/content/posts/types'
 import { imageService } from 'lib/content/posts/imageService'
 import { parseToHast } from 'lib/content/posts/parsing'
-
 import { A11yStaticImageData } from 'lib/content/images'
 import { formatDate } from 'lib/pages/posts/utils'
 import { postService } from 'lib/content/posts/postService'
@@ -44,6 +47,14 @@ export const convertRawStringToRenderablePost = async (
   const post = await postService.convertRawString(slug, rawString)
   return await convertToRenderablePost(post)
 }
+
+export const convertToReact = (
+  content: HastTree,
+  components: ComponentsWithoutNodeOptions['components']
+): React.ReactNode =>
+  unified()
+    .use(rehypeReact, { createElement, Fragment, components })
+    .stringify(content)
 
 const convertToRenderablePost = async (post: Post): Promise<RenderablePost> => {
   return {
